requirements.txt
flask>=2.0.0
pandas>=1.3.0
requests>=2.25.0
openpyxl>=3.0.0
werkzeug>=2.0.0
python-dotenv>=1.0.0
